TBR=cevans@chromium.org
Review URL: https://codereview.chromium.org/11883020

git-svn-id: svn://svn.chromium.org/chrome/branches/1364/src@176730 0039d316-1c4b-4281-b951-d872f2087c98

+++ content/renderer/pepper/pepper_plugin_delegate_impl.cc
@@ -1375,7 +1375,7 @@ PP_FlashLSORestrictions PepperPluginDelegateImpl::GetLocalDataRestrictions(
 }
 
 base::SharedMemory* PepperPluginDelegateImpl::CreateAnonymousSharedMemory(
-    uint32_t size) {
+    size_t size) {
   return RenderThread::Get()->HostAllocateSharedMemoryBuffer(size).release();
 }
